fix bug in top branch example
GitHub Actions only supports string arguments.
So the array of topBranches has to be passed as a string.
Using the Folded Block Scalar `>` seems to be the best way 
to achieve this. It makes the following like look like an array. 
And if necessary a user could split it into multiple lines.